Move types to mental-health-insights folder
